Add more action logs tests everywhere I can think of it.
