+++ tests/Feature/Checkins/Api/LicenseCheckInTest.php
@@ -4,11 +4,9 @@ namespace Tests\Feature\Checkins\Api;
 use App\Models\License;
 use App\Models\LicenseSeat;
 use App\Models\User;
-use Tests\Support\AssertsActionLogs;
 use Tests\TestCase;
 
 class LicenseCheckInTest extends TestCase {
-    use AssertsActionLogs;
     public function testLicenseCheckin()
     {
         $authUser = User::factory()->superuser()->create();
