+++ tests/Feature/Checkouts/Api/ComponentCheckoutTest.php
@@ -96,6 +96,7 @@ class ComponentCheckoutTest extends TestCase implements TestsFullMultipleCompani
             ->assertStatusMessageIs('success');
 
         $this->assertTrue($component->assets->first()->is($asset));
+        $this->assertHasTheseActionLogs($component, ['create', 'checkout']);
     }
 
     public function testComponentCheckoutIsLogged()
