+++ tests/Feature/Checkouts/Api/ConsumableCheckoutTest.php
@@ -49,6 +49,7 @@ class ConsumableCheckoutTest extends TestCase
             ]);
 
         $this->assertTrue($user->consumables->contains($consumable));
+        $this->assertHasTheseActionLogs($consumable, ['create', 'checkout']);
     }
 
     public function testUserSentNotificationUponCheckout()
