+++ tests/Feature/Checkouts/Ui/AccessoryCheckoutTest.php
@@ -83,6 +83,7 @@ class AccessoryCheckoutTest extends TestCase
             'item_type' => Accessory::class,
             'note' => 'oh hi there',
         ]);
+        $this->assertHasTheseActionLogs($accessory, ['create', 'checkout']);
     }
 
     public function testAccessoryCanBeCheckedOutWithQuantity()
@@ -109,6 +110,7 @@ class AccessoryCheckoutTest extends TestCase
             'item_type' => Accessory::class,
             'note' => 'oh hi there',
         ]);
+        $this->assertHasTheseActionLogs($accessory, ['create', 'checkout']);
     }
 
     public function testAccessoryCanBeCheckedOutToLocationWithQuantity()
@@ -135,6 +137,7 @@ class AccessoryCheckoutTest extends TestCase
             'item_type' => Accessory::class,
             'note' => 'oh hi there',
         ]);
+        $this->assertHasTheseActionLogs($accessory, ['create', 'checkout']);
     }
 
     public function testAccessoryCanBeCheckedOutToAssetWithQuantity()
@@ -161,6 +164,7 @@ class AccessoryCheckoutTest extends TestCase
             'item_type' => Accessory::class,
             'note' => 'oh hi there',
         ]);
+        $this->assertHasTheseActionLogs($accessory, ['create', 'checkout']);
     }
 
     public function testUserSentNotificationUponCheckout()
@@ -209,6 +213,7 @@ class AccessoryCheckoutTest extends TestCase
             ])->count(),
             'Log entry either does not exist or there are more than expected'
         );
+        $this->assertHasTheseActionLogs($accessory, ['create', 'checkout']);
     }
 
     public function testAccessoryCheckoutPagePostIsRedirectedIfRedirectSelectionIsIndex()
