+++ tests/Feature/Checkouts/Ui/BulkAssetCheckoutTest.php
@@ -55,6 +55,7 @@ class BulkAssetCheckoutTest extends TestCase
             $asset->assignedTo()->is($user);
             $asset->last_checkout = $checkoutAt;
             $asset->expected_checkin = $expectedCheckin;
+            $this->assertHasTheseActionLogs($asset, ['create', 'checkout']); //Note: '$this' gets auto-bound in closures, so this does work.
         });
 
         Mail::assertSent(CheckoutAssetMail::class, 2);
