+++ tests/Feature/Checkouts/Ui/ComponentsCheckoutTest.php
@@ -62,6 +62,7 @@ class ComponentsCheckoutTest extends TestCase
             ])
             ->assertStatus(302)
             ->assertRedirect(route('components.index'));
+        $this->assertHasTheseActionLogs($component, ['create', 'checkout']);
     }
 
     public function testComponentCheckoutPagePostIsRedirectedIfRedirectSelectionIsItem()
@@ -77,6 +78,7 @@ class ComponentsCheckoutTest extends TestCase
             ])
             ->assertStatus(302)
             ->assertRedirect(route('components.show', $component));
+        $this->assertHasTheseActionLogs($component, ['create', 'checkout']);
     }
 
     public function testComponentCheckoutPagePostIsRedirectedIfRedirectSelectionIsTarget()
@@ -93,5 +95,6 @@ class ComponentsCheckoutTest extends TestCase
             ])
             ->assertStatus(302)
             ->assertRedirect(route('hardware.show', $asset));
+        $this->assertHasTheseActionLogs($component, ['create', 'checkout']);
     }
 }
