+++ tests/Feature/Checkouts/Ui/ConsumableCheckoutTest.php
@@ -57,6 +57,7 @@ class ConsumableCheckoutTest extends TestCase
             ]);
 
         $this->assertTrue($user->consumables->contains($consumable));
+        $this->assertHasTheseActionLogs($consumable, ['create', 'checkout']);
     }
 
     public function testUserSentNotificationUponCheckout()
