+++ tests/Feature/Checkouts/Ui/LicenseCheckoutTest.php
@@ -39,6 +39,7 @@ class LicenseCheckoutTest extends TestCase
             'item_type' => License::class,
             'note' => 'oh hi there',
         ]);
+        $this->assertHasTheseActionLogs($licenseSeat->license, ['create', 'add seats', 'checkout']); //FIXME - out-of-order
     }
 
     public function testNotesAreStoredInActionLogOnCheckoutToUser()
@@ -62,6 +63,7 @@ class LicenseCheckoutTest extends TestCase
             'item_type' => License::class,
             'note' => 'oh hi there',
         ]);
+        $this->assertHasTheseActionLogs($licenseSeat->license, ['create', 'add seats', 'checkout']); //FIXME - out-of-order
     }
 
     public function testLicenseCheckoutPagePostIsRedirectedIfRedirectSelectionIsIndex()
