+++ tests/TestCase.php
@@ -3,8 +3,11 @@
 namespace Tests;
 
 use App\Http\Middleware\SecurityHeaders;
+use App\Models\Actionlog;
+use Illuminate\Database\Eloquent\Model;
 use Illuminate\Foundation\Testing\LazilyRefreshDatabase;
 use Illuminate\Foundation\Testing\TestCase as BaseTestCase;
+use PHPUnit\Framework\Assert;
 use RuntimeException;
 use Tests\Support\AssertsAgainstSlackNotifications;
 use Tests\Support\CanSkipTests;
@@ -46,5 +49,18 @@ abstract class TestCase extends BaseTestCase
                 '.env.testing file does not exist. Aborting to avoid wiping your local database.'
             );
         }
+    }
+
+    public function assertHasTheseActionLogs(Model $item, array $statuses)
+    {
+        \Log::error("Okay, we're running the test macro now?");
+        $logs = Actionlog::where(['item_id' => $item->id, 'item_type' => get_class($item)])->orderBy('id')->get();
+        Assert::assertEquals(count($statuses), count($logs), "Wrong count of logs expected - expecting " . count($statuses) . ", got " . count($logs));
+        $i = 0;
+        foreach ($statuses as $status) {
+            Assert::assertEquals($status, $logs[$i]->action_type, "Unexpected action type - " . $logs[$i]->action_type . " - expecting $status");
+            $i++;
+        }
+
     }
 }
